requirements.txt
openai>=2.31.0
python-dotenv>=1.0
pytest>=8.0
